pyproject.toml
[project]
name = "sparv-sbx-examples"
version = "0.1.0"
description = "Example plug-ins for Sparv)"
readme = "README.md"
license.text = "MIT License"
# authors = [
#     { name = "", email = "" }
# ]
dependencies = [
    "sparv-pipeline~=5.0"
]
entry-points."sparv.plugin" = { sbx_examples = "sbx_examples" }
